requirements.txt
asgiref==3.5.2
certifi @ file:///croot/certifi_1665076670883/work/certifi
charset-normalizer==2.1.1
Django==4.0.6
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.1
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
